+++ manga/src/main/java/io/silv/manga/domain/repositorys/SeasonalMangaRepositoryImpl.kt
@@ -3,6 +3,8 @@ package io.silv.manga.domain.repositorys
 import io.silv.core.AmadeusDispatchers
 import io.silv.ktor_response_mapper.getOrThrow
 import io.silv.manga.domain.MangaToSeasonalMangaResourceMapper
+import io.silv.manga.domain.checkProtected
+import io.silv.manga.domain.repositorys.base.LoadState
 import io.silv.manga.local.dao.SeasonalListDao
 import io.silv.manga.local.dao.SeasonalMangaResourceDao
 import io.silv.manga.local.entity.Season
@@ -13,10 +15,12 @@ import io.silv.manga.local.entity.syncerForEntity
 import io.silv.manga.network.mangadex.MangaDexApi
 import io.silv.manga.network.mangadex.models.manga.Manga
 import io.silv.manga.network.mangadex.requests.MangaRequest
+import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.onStart
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.plus
 import kotlinx.coroutines.withContext
 import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
@@ -29,6 +33,20 @@ internal class SeasonalMangaRepositoryImpl(
     private val dispatchers: AmadeusDispatchers
 ): SeasonalMangaRepository {
 
+    private val scope = CoroutineScope(dispatchers.io) + CoroutineName("SeasonalMangaRepositoryImpl")
+
+    override val loadState = MutableStateFlow<LoadState>(LoadState.None)
+
+    init {
+        scope.launch {
+            refresh()
+        }
+    }
+
+    override suspend fun refresh() {
+        getLatestLists()
+    }
+
     private fun String.toSeason() = when {
         this.contains("winter", ignoreCase = true) -> Season.Winter
         this.contains("summer", ignoreCase = true) -> Season.Summer
@@ -38,11 +56,11 @@ internal class SeasonalMangaRepositoryImpl(
     }
 
     override suspend fun updateSeasonList(season: Season, year: Int): Unit = withContext(dispatchers.io) {
+        loadState.emit(LoadState.Loading)
         runCatching {
             if (seasonalListDao.getAll().any { it.year == year  && it.season == season}) {
                 return@withContext
             }
-
             val mangaDexLists = mangaDexApi
                 .getUserLists("d2ae45e0-b5e2-4e7f-a688-17925c2d7d6b")
                 .getOrThrow()
@@ -93,33 +111,41 @@ internal class SeasonalMangaRepositoryImpl(
                     networkResponse = response
                 )
                 for (unhandled in result.unhandled) {
-                    mangaResourceDao.delete(unhandled)
+                    if (!checkProtected(unhandled.id)) {
+                        mangaResourceDao.delete(unhandled)
+                    }
                 }
             }
         }
             .onFailure { it.printStackTrace() }
+        loadState.emit(LoadState.None)
     }
 
-    override fun getSeasonalManga(id: String): Flow<SeasonalMangaResource?> {
-        return mangaResourceDao.getResourceAsFlowById(id)
+    private fun getLatestLists() = scope.launch {
+        val year = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()).year
+        Season.values().forEach {
+            updateSeasonList(
+                season = it,
+                year = year
+            )
+        }
+        Season.values().forEach {
+            updateSeasonList(
+                season = it,
+                year = year - 1
+            )
+        }
     }
 
     override fun getSeasonalLists(): Flow<List<SeasonListWithManga>> {
-        val year = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()).year
         return seasonalListDao.getSeasonListWithManga()
-            .onStart {
-                CoroutineScope(dispatchers.io).launch {
-                    Season.values().forEach {
-                        updateSeasonList(
-                            season = it,
-                            year = year
-                        )
-                        updateSeasonList(
-                            season = it,
-                            year = year - 1
-                        )
-                    }
-                }
-            }
+    }
+
+    override fun getMangaResource(id: String): Flow<SeasonalMangaResource?> {
+        return mangaResourceDao.getResourceAsFlowById(id)
+    }
+
+    override fun getAllMangaResources(): Flow<List<SeasonalMangaResource>> {
+       return mangaResourceDao.getMangaResources()
     }
 }
